Change name from faker to datagenerator
