--- hedera-mirror-faker/src/main/java/com/hedera/faker/common/TransactionGeneratorProperties.java
@@ -1,34 +1,0 @@
-package com.hedera.faker.common;
-/*-
- * ‌
- * Hedera Mirror Node
- * ​
- * Copyright (C) 2019 Hedera Hashgraph, LLC
- * ​
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * ‍
- */
-
-import javax.inject.Named;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Data
-@Named
-@RequiredArgsConstructor
-@ConfigurationProperties("faker.transaction")
-public class TransactionGeneratorProperties {
-    final private CryptoTransactionProperties crypto;
-    final private FileTransactionProperties file;
-}
