--- hedera-mirror-faker/src/main/resources/application.yml
@@ -1,49 +1,0 @@
-faker:
-  totalDuration: 1d
-  transactionsPerSecond: # 3-8 txns per second
-    rangeMin: 3
-    rangeMax: 8
-  transaction:
-    crypto:
-      frequency: 990
-      createsFrequency: 10
-      transfersFrequency: 988
-      updatesFrequency: 1
-      deletesFrequency: 1
-      numSeedAccounts: 10000
-      numTransferLists:
-        rangeMin: 1
-        rangeMax: 10
-    file:
-      frequency: 10
-      createsFrequency: 800
-      appendsFrequency: 100
-      updatesFrequency: 80
-      deletesFrequency: 20
-      numSeedFiles: 1000
-      fileDataSize:
-        rangeMin: 0
-        rangeMax: 5000
-
-logging:
-  level:
-    root: warn
-    com.hedera.faker: info
-    org.flywaydb.core.internal.command.DbMigrate: info
-
-spring:
-  datasource:
-    name: ${hedera.mirror.db.name}
-    password: ${hedera.mirror.db.password}
-    url: jdbc:postgresql://${hedera.mirror.db.host}:${hedera.mirror.db.port}/${hedera.mirror.db.name}?&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-    username: ${hedera.mirror.db.username}
-  flyway:
-    connectRetries: 10
-    baselineOnMigrate: true
-    baselineVersion: 0
-    ignoreMissingMigrations: true
-    placeholders:
-      api-password: ${hedera.mirror.db.apiPassword}
-      api-user: ${hedera.mirror.db.apiUsername}
-      db-name: ${hedera.mirror.db.name}
-      db-user: ${hedera.mirror.db.username}
